feat: Improve data persistence and UI consistency
Enhance Firestore with `experimentalForceLongPolling` and `merge: true` to mitigate connection issues and ensure data consistency across tabs. Refactor the `Avatar` component to include an `xs` size and update `App.tsx` and `types.ts` with new properties and imports for improved user data management and future features. Add a `ListIcon` to the icons component.